Fix unintended line breaks in "API access" section
The bullet points in the "API access" section are separated by newlines
in the source markdown. Unfortunately, this confuses the renderer in
such a way that in the published site the text appears *below* the
bullet point rather than next to it.

Removing the blank lines between bullet points fixes this.