Update compatibility notes on vim plugins (#76)
* Head off confusion about which vim plugin supports what

While the *hledger-vim* project has some good ideas, it is not fully
developed and hasn't been worked on for a long time. In the meanwhile
the *vim-ledger* plugin has started to support `hledger` out of the box
as one of the command backends.

At the risk of self promotion since I'm now maintaining the *vim-ledger*
plugin, this changes the order and adds a note about compatibly so it is
more apparent to a newcomer where to start and they don't skip
*vim-ledger* entirely assuming from the names that *hledger-vim* is the
only plugin that supports `hledger`.

* Fix typography, English does not allow comma before space

Leave it to a Haskell programmer to try to separate list items with
a comma on after a newline in Markdown. Note that most Markdown flavors
(including Pandoc) add a trailing space to lines when they join them.
This means you cannot start new lines with punctuation.